objekt-orientierte-scriptsprachen/photograph/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "photograph"
version = "0.1.0"
description = "Photographing / image editing software"
authors = ["Mara Schulke <mara.schulke@th-brandenburg.de>"]

[tool.poetry.scripts]
photograph = "photograph.photograph:main"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.1.0"
ttkbootstrap = "^1.10.1"
